requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.10.6
Faker==24.14.1
